+++ packages/@runlightyear/smtp/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/smtp
 
+## 0.1.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @runlightyear/lightyear@0.24.0
+
 ## 0.1.9
 
 ### Patch Changes